fix compiler warnings after stricting

+++ projecteuler/level1/p1.cc
@@ -3,7 +3,7 @@
 int main() {
     unsigned long long total = 0;
 
-    for (int i = 0; i < 1000; i++) {
+    for (unsigned long long i = 0; i < 1000; i++) {
         if (i % 3 == 0 || i % 5 == 0)
             total += i;
     }
